+++ src/graph/structs/Node.h
@@ -7,7 +7,7 @@ using namespace std;
 
 class SerializableNode : public Serializable {
 public:
-	int schemaId;
+	long schemaId;
 	SerializableNode();
 	void load(ifstream* streamObj);
 	void store(ofstream* streamObj);
@@ -16,12 +16,12 @@ public:
 class Node : public Element {
 public:
 	Schema * schema;
-	int id;
+	long id;
 	unordered_map<string, string> properties;
 	Serializable* getSerializable(string path);
 	Node();
-	Node(int id, unordered_map<string, string> properties);
-	Node(int id);
+	Node(long id, unordered_map<string, string> properties);
+	Node(long id);
 	Node(SerializableNode serializable);
 	static bool compareNodes(Node* node1, Node* node2);
 };
